Move the parameter-header support into its own mixin
The four methods that implement "Custom Headers from Tool Parameters" —
deriving the Mcp-Param-* headers, clearing that reserved namespace, fetching
the list they are derived from, and the refresh-and-retry-once recovery —
are a self-contained piece of this transport. They move to
HttpTransportBase::ParamHeaders, which keeps the module under the length
cop once the branch above this one adds to it.
