+++ lib/mcp_client/http_transport_base.rb
@@ -10,6 +10,8 @@ require_relative 'http_transport_base/sse_event_scanner'
 require_relative 'http_transport_base/stream_capture'
 require_relative 'http_transport_base/era_detection'
 
+require_relative 'http_transport_base/param_headers'
+
 module MCPClient
   # Base module for HTTP-based JSON-RPC transports
   # Contains common functionality shared between HTTP and Streamable HTTP transports
@@ -18,6 +20,7 @@ module MCPClient
     include StreamCapture
     include EraDetection
     include CalledToolDefinition
+    include ParamHeaders
 
     # Lightweight response wrapper for Faraday exception payloads (Hashes),
     # so the exception path and the default path share one challenge pipeline.
@@ -929,27 +932,6 @@ module MCPClient
       end
     end
 
-    # Attach the computed `Mcp-Param-*` headers (MCP 2026-07-28 "Custom
-    # Headers from Tool Parameters"). On a modern session that namespace is
-    # derived from the call's arguments and from nothing else -- the client
-    # MUST omit the header for an argument that is absent or null -- so a
-    # configured header of that name is cleared first: leaving it would let
-    # it stand for an argument the extraction omitted, which no tools/list
-    # refresh can correct. The clearing matches HTTP's case-insensitive field
-    # names, whatever spelling the host configured.
-    # @param req [Faraday::Request] the outgoing request
-    # @param param_headers [Hash{String => String}] the computed headers
-    # @return [void]
-    def apply_param_headers(req, param_headers)
-      if modern?
-        # The names are collected before any is dropped: the header set is
-        # being mutated.
-        configured = req.headers.keys.select { |name| MCPClient::HeaderParams.mirrored_header?(name) }
-        configured.each { |name| req.headers.delete(name) }
-      end
-      param_headers.each { |k, v| req.headers[k] = v }
-    end
-
     # Start a new session after the server invalidated the current one, then
     # resend the original request once. The @restarting_session flag prevents
     # a second restart if the fresh session also answers 404.
@@ -1040,55 +1022,6 @@ module MCPClient
       return false if sent_session_id.nil?
 
       @mutex.synchronize { !@restarting_session }
-    end
-
-    # MCP 2026-07-28 "Custom Headers from Tool Parameters": after a
-    # HeaderMismatch the client SHOULD re-fetch tools/list (the tool's
-    # inputSchema may have changed its x-mcp-header annotations) and retry
-    # the original request once with the appropriate headers. The server
-    # rejected the request before executing it, so the retry cannot
-    # duplicate a side effect. A refresh that fails re-raises the rejection:
-    # that is the actionable error.
-    # @param error [MCPClient::Errors::HeaderMismatchError] the rejection
-    # @return [void]
-    def refresh_tools_after_header_mismatch(error)
-      @logger.warn("#{sanitize_log_text(error.message)}; refreshing tools/list and retrying tools/call once")
-      refresh_tools_cache
-    rescue MCPClient::Errors::MCPError => e
-      @logger.warn("tools/list refresh after HeaderMismatch failed: #{sanitize_log_text(e.message)}")
-      raise error
-    end
-
-    # The Mcp-Param-* headers for a tools/call request (MCP 2026-07-28
-    # "Custom Headers from Tool Parameters"): the annotated arguments of the
-    # tool, looked up in this transport's tool list (fetched on demand so a
-    # call issued before tools/list still carries them).
-    # @param request [Hash] the JSON-RPC request
-    # @return [Hash{String => String}]
-    # @raise [MCPClient::Errors::ValidationError] when an annotated argument cannot be mirrored
-    def mcp_param_headers(request)
-      return {} unless request['method'] == 'tools/call'
-
-      params = request['params']
-      return {} unless params.is_a?(Hash)
-
-      name = (params['name'] || params[:name]).to_s
-      tool = known_tools_for_headers.find { |t| t.name.to_s == name }
-      # The list these headers come from is the list this request goes out
-      # under: a host re-resolving the tool after the call reads that
-      # definition back instead of asking for a possibly newer one.
-      note_called_tool_definition(name, tool)
-      return {} unless tool
-
-      MCPClient::HeaderParams.headers_for(tool.schema, params['arguments'] || params[:arguments])
-    end
-
-    # The tool list used for header extraction, fetched on demand. Mirroring
-    # is a MUST, so a list that cannot be fetched fails the call rather than
-    # letting it go out without the headers an intermediary may route on.
-    # @return [Array<MCPClient::Tool>]
-    def known_tools_for_headers
-      @mutex.synchronize { @tools } || list_tools
     end
 
     # Drop the cached tool list and re-fetch it. Hosts layered above the
